improved android support

+++ core/src/main/java/com/alibaba/fastjson2/writer/FieldWriterFloatValueFunc.java
@@ -51,9 +51,9 @@ final class FieldWriterFloatValueFunc
 
         writeFieldName(jsonWriter);
         if (decimalFormat != null) {
-            jsonWriter.writeDouble(value, decimalFormat);
+            jsonWriter.writeFloat(value, decimalFormat);
         } else {
-            jsonWriter.writeDouble(value);
+            jsonWriter.writeFloat(value);
         }
         return true;
     }
